# core-common-distributor-mgmt-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-common-distributor-mgmt
  description: Distributor Management Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/distributors
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/distributors/{distributorId}/products/{productId}:
    get:
      tags:
        - Product
      summary: Get product by ID
      description: Returns a product based on its ID
      operationId: getProductById
      parameters:
        - name: distributorId
          in: path
          description: ID of the distributor
          required: true
          schema:
            type: string
            format: uuid
        - name: productId
          in: path
          description: ID of the product to retrieve
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
        '404':
          description: Product not found
        '500':
          description: Internal server error
    put:
      tags:
        - Product
      summary: Update an existing product
      description: Updates an existing product for a specific distributor
      operationId: updateProduct
      parameters:
        - name: distributorId
          in: path
          description: ID of the distributor
          required: true
          schema:
            type: string
            format: uuid
        - name: productId
          in: path
          description: ID of the product to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDTO'
        required: true
      responses:
        '200':
          description: Product successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
        '400':
          description: Invalid product data provided
        '404':
          description: Product not found
        '500':
          description: Internal server error
    delete:
      tags:
        - Product
      summary: Delete a product
      description: Deletes a product based on its ID
      operationId: deleteProduct
      parameters:
        - name: distributorId
          in: path
          description: ID of the distributor
          required: true
          schema:
            type: string
            format: uuid
        - name: productId
          in: path
          description: ID of the product to delete
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '204':
          description: Product successfully deleted
        '404':
          description: Product not found
        '500':
          description: Internal server error
  /api/v1/distributors/{distributorId}/product-catalog/{catalogId}:
    get:
      tags:
        - Distributor Product Catalog
      summary: Get product catalog item by ID
      description: Retrieve a specific product catalog item associated with a distributor by its unique identifier
      operationId: getById
      parameters:
        - name: distributorId
          in: path
          description: Unique identifier of the distributor
          required: true
          schema:
            type: string
            format: uuid
        - name: catalogId
          in: path
          description: Unique identifier of the product catalog item
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Product catalog item found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DistributorProductCatalogDTO'
        '404':
          description: Product catalog item or distributor not found
    put:
      tags:
        - Distributor Product Catalog
      summary: Update product catalog item
      description: Update an existing product catalog item associated with a distributor
      operationId: update
      parameters:
        - name: distributorId
          in: path
          description: Unique identifier of the distributor
          required: true
          schema:
            type: string
            format: uuid
        - name: catalogId
          in: path
          description: Unique identifier of the product catalog item
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DistributorProductCatalogDTO'
        required: true
      responses:
        '200':
          description: Product catalog item successfully updated
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DistributorProductCatalogDTO'
        '400':
          description: Invalid product catalog data
        '404':
          description: Product catalog item or distributor not found
    delete:
      tags:
        - Distributor Product Catalog
      summary: Delete product catalog item
      description: Delete a product catalog item associated with a distributor from the system
      operationId: delete
      parameters:
        - name: distributorId
          in: path
          description: Unique identifier of the distributor
          required: true
          schema:
            type: string
            format: uuid
        - name: catalogId
          in: path
          description: Unique identifier of the product catalog item
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '204':
          description: Product catalog item successfully deleted
        '404':
          description: Product catalog item or distributor not found
...
  /api/v1/distributors/{distributorId}:
    get:
      tags:
        - Distributor
      summary: Get distributor by ID
      description: Returns a distributor based on its ID
      operationId: getDistributorById
      parameters:
        - name: distributorId
          in: path
          description: ID of the distributor to retrieve
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved distributor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistributorDTO'
        '404':
          description: Distributor not found
        '500':
          description: Internal server error
    put:
      tags:
        - Distributor
      summary: Update distributor
      description: Updates an existing distributor with the provided information
      operationId: updateDistributor
      parameters:
        - name: distributorId
          in: path
          description: ID of the distributor to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DistributorDTO'
        required: true
      responses:
        '200':
          description: Distributor successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistributorDTO'
        '400':
          description: Invalid distributor data provided
        '404':
          description: Distributor not found
        '500':
          description: Internal server error
    delete:
      tags:
        - Distributor
      summary: Delete distributor
      description: Deletes a distributor based on its ID
      operationId: deleteDistributor
      parameters:
        - name: distributorId
          in: path
          description: ID of the distributor to delete
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '204':
          description: Distributor successfully deleted
        '404':
          description: Distributor not found
        '500':
          description: Internal server error
...
  /api/v1/distributors/{distributorId}/products/filter:
    post:
      tags:
        - Product
      summary: Filter products
      description: Returns a paginated list of products based on filter criteria for a specific distributor
      operationId: filterProducts
      parameters:
        - name: distributorId
          in: path
          description: ID of the distributor
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestProductDTO'
        required: true
      responses:
        '200':
          description: Successfully retrieved products
          content:
            application/json: {}
        '400':
          description: Invalid filter criteria provided
        '500':
          description: Internal server error
...
  /api/v1/distributors/{distributorId}/product-catalog:
    post:
      tags:
        - Distributor Product Catalog
      summary: Create product catalog item for a distributor
      description: Create a new product catalog item associated with a specific distributor
      operationId: create
      parameters:
        - name: distributorId
          in: path
          description: Unique identifier of the distributor
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DistributorProductCatalogDTO'
        required: true
      responses:
        '201':
          description: Product catalog item successfully created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DistributorProductCatalogDTO'
        '400':
          description: Invalid product catalog data
        '404':
          description: Distributor not found
  /api/v1/distributors/{distributorId}/product-catalog/filter:
    post:
      tags:
        - Distributor Product Catalog
      summary: Filter product catalog for a distributor
      description: Retrieve a paginated list of product catalog items associated with a specific distributor based on filtering criteria
      operationId: filter
      parameters:
        - name: distributorId
          in: path
          description: Unique identifier of the distributor
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestDistributorProductCatalogDTO'
        required: true
      responses:
        '200':
          description: Successfully retrieved filtered product catalog
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '400':
          description: Invalid filter request
        '404':
          description: Distributor not found
...
  /api/v1/distributors/filter:
    post:
      tags:
        - Distributor
      summary: Filter distributors
      description: Returns a paginated list of distributors based on filter criteria
      operationId: filterDistributors
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestDistributorDTO'
        required: true
      responses:
        '200':
          description: Successfully retrieved distributors
          content:
            application/json: {}
        '400':
          description: Invalid filter criteria provided
        '500':
          description: Internal server error
...
  /api/v1/shipments/product/{productId}:
    get:
      tags:
        - Shipment
      summary: Get shipments by product
      description: Returns all shipments associated with a specific product
      operationId: getShipmentsByProductId
      parameters:
        - name: productId
          in: path
          description: ID of the product
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved shipments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShipmentDTO'
        '404':
          description: Product not found
        '500':
          description: Internal server error
...
  /api/v1/lending-contracts/product/{productId}:
    get:
      tags:
        - Lending Contract
      summary: Get lending contracts by product
      description: Returns all lending contracts associated with a specific product
      operationId: getLendingContractsByProductId
      parameters:
        - name: productId
          in: path
          description: ID of the product
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved lending contracts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LendingContractDTO'
        '404':
          description: Product not found
        '500':
          description: Internal server error
...
  /api/v1/lending-contracts/distributor/{distributorId}:
    get:
      tags:
        - Lending Contract
      summary: Get lending contracts by distributor
      description: Returns all lending contracts associated with a specific distributor
      operationId: getLendingContractsByDistributorId
      parameters:
        - name: distributorId
          in: path
          description: ID of the distributor
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved lending contracts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LendingContractDTO'
        '404':
          description: Distributor not found
        '500':
          description: Internal server error
...
components:
  schemas:
...
    DistributorDTO:
      required:
        - name
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        externalCode:
          maxLength: 255
          minLength: 0
          type: string
        name:
          maxLength: 255
          minLength: 0
          type: string
        displayName:
          maxLength: 255
          minLength: 0
          type: string
        taxId:
          type: string
        registrationNumber:
          maxLength: 100
          minLength: 0
          type: string
        websiteUrl:
          maxLength: 255
          minLength: 0
          type: string
        phoneNumber:
          type: string
        email:
          type: string
        supportEmail:
          type: string
        addressLine:
          maxLength: 255
          minLength: 0
          type: string
        postalCode:
          maxLength: 20
          minLength: 0
          type: string
        city:
          maxLength: 100
          minLength: 0
          type: string
        state:
          maxLength: 100
          minLength: 0
          type: string
        countryId:
          type: string
          format: uuid
        isActive:
          type: boolean
        isTestDistributor:
          type: boolean
        timeZone:
          maxLength: 50
          minLength: 0
          type: string
        defaultLocale:
          maxLength: 10
          minLength: 0
          type: string
        onboardedAt:
          type: string
          format: date-time
        terminatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          format: uuid
...
    ProductDTO:
      required:
        - distributorId
        - name
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        distributorId:
          type: string
          format: uuid
        name:
          maxLength: 255
          minLength: 0
          type: string
        description:
          maxLength: 1000
          minLength: 0
          type: string
        sku:
          maxLength: 100
          minLength: 0
          type: string
        modelNumber:
          maxLength: 100
          minLength: 0
          type: string
        manufacturer:
          maxLength: 255
          minLength: 0
          type: string
        categoryId:
          type: string
          format: uuid
        imageUrl:
          maxLength: 255
          minLength: 0
          type: string
        specifications:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          format: uuid
    DistributorProductCatalogDTO:
      required:
        - distributorId
        - productId
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        distributorId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        catalogCode:
          type: string
        displayName:
          type: string
        customDescription:
          type: string
        isFeatured:
          type: boolean
        isAvailable:
          type: boolean
        availabilityStartDate:
          type: string
          format: date-time
        availabilityEndDate:
          type: string
          format: date-time
        displayOrder:
          type: integer
          format: int32
        minQuantity:
          type: integer
          format: int32
        maxQuantity:
          type: integer
          format: int32
        shippingAvailable:
          type: boolean
        shippingCost:
          type: number
        shippingTimeDays:
          type: integer
          format: int32
        specialConditions:
          type: string
        metadata:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
          readOnly: true
        createdBy:
          type: string
          format: uuid
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
        updatedBy:
          type: string
          format: uuid
          readOnly: true
...
    FilterOptions:
      type: object
      properties:
        caseInsensitiveStrings:
          type: boolean
        includeInheritedFields:
          type: boolean
      description: Filter options for customizing filter behavior
...
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: The direction of sorting, either ASC or DESC.
          example: DESC
          default: DESC
      description: Represents a pagination request for retrieving paginated results, including page number, size, sort field, and direction.
    RangeFilter:
      type: object
      properties:
        ranges:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/RangeObject'
          description: Map of field names to their range criteria
      description: Represents range filters for numeric fields
...
    FilterRequestProductDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/ProductDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Generic filter request that includes both filter criteria and pagination
    FilterRequestDistributorProductCatalogDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/DistributorProductCatalogDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Filter criteria for product catalog
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...
    FilterRequestDistributorDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/DistributorDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Generic filter request that includes both filter criteria and pagination
...